Fix CI linting failures and consolidate coverage tests.
- Fixed SIM108 (ternary operator) in builder.py.
- Fixed PT011 (broad exception match) in test assertions.
- Distributed coverage tests from test_coverage_final.py to relevant test files to maintain 100% coverage cleanly.
- Verified 100% coverage locally.
