+++ src/coreason_manifest/spec/core/flow.py
@@ -1,4 +1,3 @@
-import warnings
 from collections.abc import Iterable
 from typing import Annotated, Any, Literal
